Bugfixes - Fix how Point is serialized - Reconnect within an interval - Use UInt64 instead of Long - Some refactoring

+++ app/src/main/java/com/lightning/wallet/test/WireSpec.scala
@@ -8,6 +8,7 @@ import com.lightning.wallet.ln.wire.LightningMessageCodecs._
 import scodec.bits.{BitVector, ByteVector, HexStringSyntax}
 import fr.acinq.bitcoin.{BinaryData, Crypto}
 import fr.acinq.bitcoin.Crypto.{PrivateKey, PublicKey, Scalar}
+import fr.acinq.eclair.UInt64
 
 import scala.util.Random
 
@@ -141,8 +142,8 @@ class WireSpec {
     {
       println("encode/decode all channel messages")
 
-      val open = OpenChannel(randomBytes(32), randomBytes(32), 3, 4, 5, 6, 7, 8, 9, 10, 11, publicKey(1), point(2), point(3), point(4), point(5), 0.toByte)
-      val accept = AcceptChannel(randomBytes(32), 3, 4, 5, 6, 7, 8, 9, publicKey(1), point(2), point(3), point(4), point(5))
+      val open = OpenChannel(randomBytes(32), randomBytes(32), 3, 4, 5, UInt64(6), 7, 8, 9, 10, 11, publicKey(1), point(2), point(3), point(4), point(5), 0.toByte)
+      val accept = AcceptChannel(randomBytes(32), 3, UInt64(4), 5, 6, 7, 8, 9, publicKey(1), point(2), point(3), point(4), point(5))
       val funding_created = FundingCreated(randomBytes(32), bin(32, 0), 3, randomSignature)
       val funding_signed = FundingSigned(randomBytes(32), randomSignature)
       val funding_locked = FundingLocked(randomBytes(32), point(2))
